Inital tab update
Added a tab functionality to the filter activity. This is to separate
"place" and "injury " filtering. Not working currently...
