+++ app/src/main/java/com/lol2kpe/h4u/filter/SymptomFragment.java
@@ -1,0 +1,20 @@
+package com.lol2kpe.h4u.filter;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.lol2kpe.h4u.R;
+
+/**
+ * Created by Jonathan on 2017-04-24.
+ */
+public class SymptomFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        return inflater.inflate(R.layout.symptom_tab, container, false);
+    }
+}
